make test behavior a bit more configurable

+++ src/test/kotlin/io/inbot/search/escrud/AbstractElasticSearchTest.kt
@@ -8,7 +8,7 @@ import org.elasticsearch.client.RestHighLevelClient
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 
-open class AbstractElasticSearchTest {
+open class AbstractElasticSearchTest(val indexPrefix: String = "test", val deleteIndexAfterTest: Boolean = true) {
     lateinit var dao: IndexDAO<TestModel>
     lateinit var esClient: RestHighLevelClient
     lateinit var indexName: String
@@ -18,13 +18,15 @@ open class AbstractElasticSearchTest {
         val restClientBuilder = RestClient.builder(HttpHost("localhost", 9999, "http"))
         esClient = RestHighLevelClient(restClientBuilder)
         // each test gets a fresh index
-        indexName = "test-" + randomId()
+        indexName = "$indexPrefix-" + randomId()
         dao = esClient.crudDao(indexName, refreshAllowed = true, modelReaderAndWriter = JacksonModelReaderAndWriter(TestModel::class, ObjectMapper().findAndRegisterModules()))
     }
 
     @AfterEach
     fun after() {
         // delete the index after the test
-        esClient.indices().delete(DeleteIndexRequest(indexName))
+        if (deleteIndexAfterTest) {
+            esClient.indices().delete(DeleteIndexRequest(indexName))
+        }
     }
 }
